requirements.txt
python-dotenv==1.2.1
streamlit==1.41.1
langchain==0.3.26
openai==1.91.0
langchain-community==0.3.26
langchain-openai==0.3.27